Gerado MessagesController e movido endpoint Messages/Register
